chore: refactor other identity plugins
