+++ src/plugins/guestIdentity/plugin.ts
@@ -4,6 +4,6 @@ import { GuestIdentityDriver } from './GuestIdentityDriver';
 
 export const guestIdentity = (): MetaplexPlugin => ({
   install(metaplex: Metaplex) {
-    metaplex.setIdentityDriver(new GuestIdentityDriver(metaplex));
+    metaplex.identity().setDriver(new GuestIdentityDriver());
   },
 });
