+++ src/plugins/keypairIdentity/KeypairIdentityDriver.ts
@@ -1,18 +1,14 @@
 import { Keypair, PublicKey, Transaction } from '@solana/web3.js';
 import nacl from 'tweetnacl';
-import { Metaplex } from '@/Metaplex';
-import { IdentityDriver, KeypairSigner } from '@/types';
+import { KeypairSigner } from '@/types';
+import { IdentityDriver } from '../identityModule';
 
-export class KeypairIdentityDriver
-  extends IdentityDriver
-  implements KeypairSigner
-{
+export class KeypairIdentityDriver implements IdentityDriver, KeypairSigner {
   public readonly keypair: Keypair;
   public readonly publicKey: PublicKey;
   public readonly secretKey: Uint8Array;
 
-  constructor(metaplex: Metaplex, keypair: Keypair) {
-    super(metaplex);
+  constructor(keypair: Keypair) {
     this.keypair = keypair;
     this.publicKey = keypair.publicKey;
     this.secretKey = keypair.secretKey;
